Backport cc8bf95046d1fba0f88b0e17481f36b2be870659
